kubernetes-deployments-adilramzan/app/requirements.txt
flask==2.0.1
werkzeug==2.0.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
